Add Gradle support.
