--- RealStoragePathLibrary-Example/lint.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<lint>
-    <issue id="NewApi" severity="ignore">
-        <ignore path="src/com/sandisk/realstoragepath/RealStoragePath.java" />
-    </issue>
-</lint>
