+++ library/src/main/AndroidManifest.xml
@@ -1,0 +1,8 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+  package="com.sandisk.realstoragepath">
+
+  <application android:allowBackup="true" android:label="@string/app_name">
+
+  </application>
+
+</manifest>
